Adicionado metodo Lab

+++ Program.cs
@@ -3,12 +3,10 @@ using LabManager.Database;
 using LabManager.Repositories;
 using LabManager.Models;
 
-
 var databaseConfig = new DatabaseConfig();
-
 var databaseSetup = new DatabaseSetup(databaseConfig);
-
 var computerRepository = new ComputerRepository(databaseConfig);
+var labRepository = new LabRepository(databaseConfig);
 
 var modelName = args[0];
 var modelAction = args[1];
@@ -23,6 +21,7 @@ if(modelName == "Computer")
             Console.WriteLine($"{computer.Id}, {computer.Ram}, {computer.Processor}");
         }
     }
+
     if(modelAction == "New")
     {
         var id = Convert.ToInt32(args[2]);
@@ -33,82 +32,93 @@ if(modelName == "Computer")
         computerRepository.Save(computer);
     }
 
- if(modelAction == "Show")
+    if(modelAction == "Show")
     {
         var id = Convert.ToInt32(args[2]);
         if(computerRepository.ExistsById(id))
-
         {
-           var computer = computerRepository.GetById(id);
-           Console.WriteLine($"{computer.Id}, {computer.Ram}, {computer.Processor}");
+        var computer = computerRepository.GetById(id);
+        Console.WriteLine($"{computer.Id}, {computer.Ram}, {computer.Processor}");
         } else {
             Console.WriteLine($"O computador com id {id} não existe");
         }
- 
     }
 
     if(modelAction == "Update")
     {
         var id = Convert.ToInt32(args[2]);
+        if(computerRepository.ExistsById(id))
+        {
         string ram = args[3];
         string processor = args[4];
 
         var computer = new Computer(id, ram, processor);
         computerRepository.Update(computer);
+        } else {
+            Console.WriteLine($"O computador com id {id} não existe");
+        }
     }
-
-
-if(modelAction == "Delete")
+    
+    if(modelAction == "Delete")
     {
         var id = Convert.ToInt32(args[2]);
+        if(computerRepository.ExistsById(id))
+        {
         computerRepository.Delete(id);
+        } else {
+            Console.WriteLine($"O computador com id {id} não existe");
+        }
     }
-
-
 }
 if(modelName == "Lab")
 {
     if(modelAction == "List")
     {
-        Console.WriteLine("Lab List");
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Lab";
-        var reader = command.ExecuteReader();
-        
-        while(reader.Read())
+        Console.WriteLine("Lista de Laboratórios:");
+        foreach (var lab in labRepository.GetAll())
         {
-            Console.WriteLine("{0}, {1}, {2}, {3}", reader.GetInt32(0), reader.GetInt32(1), reader.GetString(2), reader.GetString(3));
+            Console.WriteLine($"{lab.Id}, {lab.Number}, {lab.Name}, {lab.Block}");
         }
-        
-        connection.Close();
     }
-    
+
     if(modelAction == "New")
     {
         var id = Convert.ToInt32(args[2]);
         var number = args[3];
         string name = args[4];
         string block = args[5];
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
-        var command = connection.CreateCommand();
-        command.CommandText = "INSERT INTO Lab VALUES($id, $number, $name, $block)";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$number", number);
-        command.Parameters.AddWithValue("$name", name);
-        command.Parameters.AddWithValue("$block", block);
 
-        command.ExecuteNonQuery();
-
-        connection.Close();
-
-
+        var lab = new Lab(id, number, name, block);
+        labRepository.Save(lab);
     }
 
-  
+    if(modelAction == "Update")
+    {
+        var id = Convert.ToInt32(args[2]);
+        var number = args[3];
+        string name = args[4];
+        string block = args[5];
 
+        var lab = new Lab(id, number, name, block);
+        labRepository.Update(lab);
+    }
 
+    if(modelAction == "Show")
+    {
+        var id = Convert.ToInt32(args[2]);
+        if(labRepository.ExistsById(id))
+        {
+            var lab = labRepository.GetById(id);
+            Console.WriteLine($"{lab.Id}, {lab.Number}, {lab.Name}, {lab.Block}");
+        } else{
+            Console.WriteLine($"O Lab com Id {id} não existe");
+        }
+    
+    }
 
+    if(modelAction == "Delete")
+    {
+        var id = Convert.ToInt32(args[2]);
+        labRepository.Delete(id);
+    }
 }
